Get Data from API

+++ app/src/main/java/com/devotion/makancuy/data/repository/CartRepository.kt
@@ -17,8 +17,8 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onStart
 
 interface CartRepository {
-    fun getUserCartData(): Flow<ResultWrapper<Pair<List<Cart>, Double>>>
-    fun getCheckoutData(): Flow<ResultWrapper<Triple<List<Cart>,List<PriceItem>, Double>>>
+    fun getUserCartData(): Flow<ResultWrapper<Pair<List<Cart>, Int>>>
+    fun getCheckoutData(): Flow<ResultWrapper<Triple<List<Cart>,List<PriceItem>, Int>>>
 
     fun createCart(
         menu : Menu,
@@ -33,7 +33,7 @@ interface CartRepository {
 }
 
 class CartRepositoryImpl(private val cartDataSource : CartDataSource) : CartRepository{
-    override fun getUserCartData(): Flow<ResultWrapper<Pair<List<Cart>, Double>>> {
+    override fun getUserCartData(): Flow<ResultWrapper<Pair<List<Cart>, Int>>> {
         return cartDataSource.getAllCarts()
             .map {
                 //mapping into cart list and sum the total price
@@ -52,7 +52,7 @@ class CartRepositoryImpl(private val cartDataSource : CartDataSource) : CartRepo
             }
     }
 
-    override fun getCheckoutData(): Flow<ResultWrapper<Triple<List<Cart>, List<PriceItem>, Double>>> {
+    override fun getCheckoutData(): Flow<ResultWrapper<Triple<List<Cart>, List<PriceItem>, Int>>> {
         return cartDataSource.getAllCarts()
             .map {
                 proceed {
